Add fluent validation to user controllers

+++ CustomEd.User.Service/src/Controllers/StudentController.cs
@@ -2,6 +2,9 @@ using Microsoft.AspNetCore.Mvc;
 using CustomEd.User.Service.Model;
 using CustomEd.User.Service.Data.Interfaces;
 using CustomEd.User.Service.Response;
+using CustomEd.User.Service.DTOs;
+using AutoMapper;
+using CustomEd.User.Service.Validators;
 
 namespace CustomEd.User.Service.Controllers
 {
@@ -9,47 +12,72 @@ namespace CustomEd.User.Service.Controllers
     [Route("api/user/student")]
     public class StudentController : UserController<Model.Student>
     {
-        public StudentController(IGenericRepository<Model.Student> userRepository) : base(userRepository)
+        public StudentController(IGenericRepository<Model.Student> userRepository, IMapper mapper) : base(userRepository, mapper)
         {
         }
 
         [HttpGet("student-id")]
-        public async Task<ActionResult<SharedResponse<Model.Student>>> SearchStudentBySchoolId([FromQuery] string id)
+        public async Task<ActionResult<SharedResponse<StudentDto>>> SearchStudentBySchoolId([FromQuery] string id)
         {
             var student = await _userRepository.GetAsync(u => u.StudentId == id);
-            return Ok(SharedResponse<Model.Student>.Success(student, "Students retrieved successfully"));
+            var studentDto = _mapper.Map<StudentDto>(student);
+            return Ok(SharedResponse<StudentDto>.Success(studentDto, "Students retrieved successfully"));
         }
 
         [HttpGet("student-name")]
-        public async Task<ActionResult<SharedResponse<IEnumerable<Model.Student>>>> SearchStudentByName([FromQuery] string name)
+        public async Task<ActionResult<SharedResponse<IEnumerable<StudentDto>>>> SearchStudentByName([FromQuery] string name)
         {
             var students = await _userRepository.GetAllAsync(u => u.FirstName!.Contains(name) || u.LastName!.Contains(name));
-            return Ok(SharedResponse<IEnumerable<Model.Student>>.Success(students, "Students retrieved successfully"));
+            var studentsDto = _mapper.Map<IEnumerable<StudentDto>>(students);
+            return Ok(SharedResponse<IEnumerable<StudentDto>>.Success(studentsDto, "Students retrieved successfully"));
         }
 
         [HttpPost]
-        public async Task<ActionResult<SharedResponse<Model.Student>>> CreateUser([FromBody] Model.Student user)
+        public async Task<ActionResult<SharedResponse<Model.Student>>> CreateUser([FromBody] CreateStudentDto studentDto)
         {
-            await _userRepository.CreateAsync(user);
-            return CreatedAtAction(nameof(GetUserById), new { id = user.Id }, SharedResponse<Model.Student>.Success(user, "User created successfully"));
+
+            var createStudentDtoValidator = new CreateStudentDtoValidator();
+            var validationResult = await createStudentDtoValidator.ValidateAsync(studentDto);
+            if (!validationResult.IsValid)
+            {
+                return BadRequest(SharedResponse<Model.Student>.Fail("Invalid input", validationResult.Errors.Select(e => e.ErrorMessage).ToList()));
+            }
+
+            var student = _mapper.Map<Model.Student>(studentDto);
+            
+            await _userRepository.CreateAsync(student);
+            return CreatedAtAction(nameof(GetUserById), new { id = student.Id }, SharedResponse<Model.Student>.Success(student, "User created successfully"));
             
         }
 
-        [HttpDelete]
-        public async Task<ActionResult<SharedResponse<Model.Student>>> RemoveUser(Guid id)
+        [HttpDelete("{id}")]
+        public async Task<ActionResult<SharedResponse<StudentDto>>> RemoveUser(Guid id)
         {
+            if (id == Guid.Empty || await _userRepository.GetAsync(id) == null)
+            {
+                return BadRequest(SharedResponse<StudentDto>.Fail("Invalid Id", new List<string> { "Invalid id" }));
+            }
+
+
             await _userRepository.RemoveAsync(id);
-            return Ok(SharedResponse<Model.Student>.Success(null, "User deleted successfully"));
+            return Ok(SharedResponse<StudentDto>.Success(null, "User deleted successfully"));
         }
 
         [HttpPut]
-        public async Task<ActionResult<SharedResponse<Model.Student>>> UpdateUser([FromBody] Model.Student user)
+        public async Task<ActionResult<SharedResponse<StudentDto>>> UpdateUser([FromBody] UpdateStudentDto studentDto)
         {
-            await _userRepository.UpdateAsync(user);
-            return Ok(SharedResponse<Model.Student>.Success(null, "User updated successfully"));
-            
-        } 
+            var updateStudentDtoValidator = new UpdateStudentDtoValidator(_userRepository);
+            var validationResult = await updateStudentDtoValidator.ValidateAsync(studentDto);
+            if (!validationResult.IsValid)
+            {
+                return BadRequest(SharedResponse<StudentDto>.Fail("Invalid input", validationResult.Errors.Select(e => e.ErrorMessage).ToList()));
+            }
 
+            var student = _mapper.Map<Model.Student>(studentDto);
+            await _userRepository.UpdateAsync(student);
+            return Ok(SharedResponse<StudentDto>.Success(null, "User updated successfully"));
+            
+        }
         
     }
 }
